+++ include/mpi/core/communicators/graph_communicator.hpp
@@ -24,7 +24,7 @@ public:
   {
     managed_ = true;
 
-    MPI_CHECK_ERROR_CODE(MPI_Graph_create, (that.native(), static_cast<std::int32_t>(graph.degrees.size()), graph.degrees.data(), graph.edges.data(), static_cast<std::int32_t>(reorder), &native_))
+    MPI_CHECK_ERROR_CODE(MPI_Graph_create, (that.native(), static_cast<std::int32_t>(graph.degrees.size()), graph.degrees.data(), graph.edges.data(), reorder, &native_))
   }
   graph_communicator            (const graph_communicator&  that)          = default;
   graph_communicator            (      graph_communicator&& temp) noexcept = default;
